pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "reciply"
version = "0"
requires-python = ">= 3.11"

# ----------
# Dependencies
# ----------

dependencies = [
    "attrs",
    "django",
    "django-configurations",
    "psycopg[binary]",
    "python-decouple",
    "pulp"
]

[project.optional-dependencies]

ci = [
    # Test framework
    "pytest",
    "pytest-django",
    # Test setup
    "factory_boy",
    # Functional testing
    "django-webtest",
    "webtest",
    # Linting / typing
    "django-stubs",
    "flake8",
    "ruff",
    "isort",
    "mypy",
    # Needs relegating to `dev`
    "django-extensions",
]

dev = [
    # Git
    "pre-commit",
    # Pytest plugins for assisting local dev (but not used in CI)
    "pytest-random-order",
    "pytest-repeat",
]

# ----------
# Tooling
# ----------

[tool.pytest.ini_options]
pythonpath = "./src/reciply"
DJANGO_SETTINGS_MODULE = "reciply.config.settings"
DJANGO_CONFIGURATION = "Settings"

filterwarnings = [
    "ignore:The FORMS_URLFIELD_ASSUME_HTTPS transitional setting is deprecated:django.utils.deprecation.RemovedInDjango60Warning:django",
]

[tool.mypy]
ignore_missing_imports = true
exclude = ["venv"]
plugins = ["./init_mypy.py"]

[[tool.mypy.overrides]]
module = "tests.*"
allow_untyped_defs = true

[tool.django-stubs]
django_settings_module = "reciply.config.settings"


[tool.ruff]

# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint]
select = [
    "I", # isort
]

[tool.ruff.lint.isort]
# Sections
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.per-file-ignores]
"*__init__.py" = ["F401"]  # Allow 'unused' imports in __init__.py files.
